Desktop/upwork/3) current projects/solotrendx_021891146667638771828/requirements.txt
# Core requirements
flask==3.1.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.30.0
PyJWT>=2.6.0
flask-cors>=3.0.10

# Testing requirements
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0  # For mocking HTTP requests

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0